pyproject.toml
[project]
name = "paddling"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13,<3.14"
dependencies = [
    "alembic==1.13.1",
    "blinker==1.9.0",
    "certifi==2025.4.26",
    "charset-normalizer==3.4.2",
    "click==8.2.1",
    "deprecated==1.2.18",
    "flask==3.1.3",
    "flask-limiter==3.12",
    "flask-login==0.6.3",
    "flask-sqlalchemy==3.1.1",
    "flask-wtf==1.2.2",
    "greenlet==3.2.2",
    "idna==3.10",
    "itsdangerous==2.2.0",
    "jinja2==3.1.6",
    "limits==5.2.0",
    "markdown-it-py==3.0.0",
    "markupsafe==3.0.2",
    "mdurl==0.1.2",
    "ordered-set==4.1.0",
    "packaging==25.0",
    "psycopg[binary]==3.2.9",
    "pydocstyle==6.3.0",
    "pygments==2.19.1",
    "pytest==8.2.0",
    "python-dotenv==1.1.0",
    "requests==2.32.4",
    "rich==13.9.4",
    "sqlalchemy==2.0.41",
    "typing-extensions==4.13.2",
    "urllib3==2.6.3",
    "werkzeug==3.1.6",
    "wrapt==1.17.2",
    "wtforms==3.2.1",
]

[dependency-groups]
dev = [
    "bandit>=1.8.6",
    "black>=26.3.1",
    "coverage>=7.10.6",
    "filelock>=3.20.3",
    "mypy>=1.17.1",
    "pillow>=11.3.0",
    "pip>=25.3",
    "pip-audit>=2.9.0",
    "pytest>=8.2.0",
    "ruff>=0.12.11",
    "types-requests>=2.32.0",
]


# You can also be explicit instead of `strict` and toggle rules you want.

[[tool.mypy.overrides]]
module = ["flask_login.*", "flask_wtf.*"]
ignore_missing_imports = true

[tool.black]
target-version = ["py313"]
